rename: AuthorityStatus -> enumerate 패키지로 이동 및 seller/constant 패키지 삭제
