+++ src/main/java/org/boot/growup/common/enumerate/AuthorityStatus.java
@@ -1,4 +1,4 @@
-package org.boot.growup.source.seller.constant;
+package org.boot.growup.common.enumerate;
 
 public enum AuthorityStatus {
     APPROVED, DENIED, PENDING
